+++ baseapp/baseapp.go
@@ -246,18 +246,18 @@ func UpgradeableStoreLoader(upgradeInfoPath string) StoreLoader {
 		// there is a migration file, let's execute
 		data, err := ioutil.ReadFile(upgradeInfoPath)
 		if err != nil {
-			return fmt.Errorf("Cannot read upgrade file %s: %v", upgradeInfoPath, err)
+			return fmt.Errorf("cannot read upgrade file %s: %v", upgradeInfoPath, err)
 		}
 
 		var upgrades storetypes.StoreUpgrades
 		err = json.Unmarshal(data, &upgrades)
 		if err != nil {
-			return fmt.Errorf("Cannot parse upgrade file: %v", err)
+			return fmt.Errorf("cannot parse upgrade file: %v", err)
 		}
 
 		err = ms.LoadLatestVersionAndUpgrade(&upgrades)
 		if err != nil {
-			return fmt.Errorf("Load and upgrade database: %v", err)
+			return fmt.Errorf("load and upgrade database: %v", err)
 		}
 
 		// if we have a successful load, we delete the file