+++ server/start.go
@@ -99,6 +99,7 @@ func startStandAlone(ctx *Context, appCreator AppCreator) error {
 
 	// run forever (the node will not be returned)
 	select {}
+
 }
 
 func startInProcess(ctx *Context, appCreator AppCreator) (*node.Node, error) {