+++ CHANGELOG.md
@@ -65,6 +65,7 @@ longer panics if the store to load contains substores that we didn't explicitly 
   * Implement `SimulationManager` for executing modules' simulation functionalities in a modularized way
   * Add `DecodeStore` to the `SimulationManager` for decoding each module's types
 * (simulation) [\#4893](https://github.com/cosmos/cosmos-sdk/issues/4893) Change SimApp keepers to be public and add getter functions for keys and codec
+* (simulation) [\#4906](https://github.com/cosmos/cosmos-sdk/issues/4906) Add simulation `Config` struct that wraps simulation flags
 * (store) [\#4792](https://github.com/cosmos/cosmos-sdk/issues/4792) panic on non-registered store
 * (types) [\#4821](https://github.com/cosmos/cosmos-sdk/issues/4821) types/errors package added with support for stacktraces. It is meant as a more feature-rich replacement for sdk.Errors in the mid-term.
 
@@ -76,6 +77,7 @@ longer panics if the store to load contains substores that we didn't explicitly 
   * Update `handleQueryStore` to resemble `handleQueryCustom`
 * (cli) [\#4763](https://github.com/cosmos/cosmos-sdk/issues/4763) Fix flag `--min-self-delegation` for staking `EditValidator`
 * (keys) Fix ledger custom coin type support bug
+* (simulation) [\#4912](https://github.com/cosmos/cosmos-sdk/issues/4912) Fix SimApp ModuleAccountAddrs to properly return black listed addresses for bank keeper initialization
 
 ## [v0.36.0] - 2019-08-13